refactor(electron): split oversized advisor adapters
